Update java packages from `anthos.samples.financedemo` to `anthos.samples.bankofanthos` in balance reader, transaction history and ledger writer modules
