+++ src/balancereader/src/main/java/anthos/samples/bankofanthos/balancereader/BalanceReaderController.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package anthos.samples.financedemo.transactionhistory;
+package anthos.samples.bankofanthos.balancereader;
 
 import java.io.IOException;
 
@@ -28,17 +28,12 @@ import java.security.spec.InvalidKeySpecException;
 import java.security.spec.X509EncodedKeySpec;
 
 import java.util.Base64;
-import java.util.Collection;
-import java.util.Deque;
 import java.util.concurrent.ExecutionException;
-import java.util.concurrent.TimeUnit;
 import java.util.logging.Logger;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.dao.DataAccessResourceFailureException;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -60,29 +55,25 @@ import com.google.common.cache.LoadingCache;
 import com.google.common.util.concurrent.UncheckedExecutionException;
 
 /**
- * Controller for the TransactionHistory service.
- *
- * Functions to show the transaction history for each user account.
+ * REST service to retrieve the current balance for the authenticated user.
  */
 @RestController
-public final class TransactionHistoryController {
+public final class BalanceReaderController {
 
     private static final Logger LOGGER =
-            Logger.getLogger(TransactionHistoryController.class.getName());
+            Logger.getLogger(BalanceReaderController.class.getName());
 
     @Autowired
     private TransactionRepository dbRepo;
 
-    @Value("${EXTRA_LATENCY_MILLIS:#{null}}")
-    private Integer extraLatencyMillis;
-    @Value("${HISTORY_LIMIT:100}")
-    private Integer historyLimit;
+    @Value("${LOCAL_ROUTING_NUM}")
+    private String localRoutingNum;
     @Value("${VERSION}")
     private String version;
 
     private JWTVerifier verifier;
+    private LoadingCache<String, Long> cache;
     private LedgerReader ledgerReader;
-    private LoadingCache<String, Deque<Transaction>> cache;
 
     /**
      * Constructor.
@@ -90,10 +81,9 @@ public final class TransactionHistoryController {
      * Initializes JWT verifier and a connection to the bank ledger.
      */
     @Autowired
-    public TransactionHistoryController(LedgerReader reader,
+    public BalanceReaderController(LedgerReader reader,
             @Value("${PUB_KEY_PATH}") final String publicKeyPath,
-            @Value("${CACHE_SIZE:1000}") final Integer expireSize,
-            @Value("${CACHE_MINUTES:60}") final Integer expireMinutes,
+            @Value("${CACHE_SIZE:1000000}") final Integer expireSize,
             @Value("${LOCAL_ROUTING_NUM}") final String localRoutingNum) {
         // Initialize JWT verifier.
         try {
@@ -110,28 +100,28 @@ public final class TransactionHistoryController {
             Algorithm algorithm = Algorithm.RSA256(publicKey, null);
             this.verifier = JWT.require(algorithm).build();
         } catch (IOException
-                 | NoSuchAlgorithmException
-                 | InvalidKeySpecException e) {
+                | NoSuchAlgorithmException
+                | InvalidKeySpecException e) {
             LOGGER.severe(e.toString());
             System.exit(1);
         }
         // Initialize cache
-        CacheLoader load = new CacheLoader<String, Deque<Transaction>>() {
+        CacheLoader loader =  new CacheLoader<String, Long>() {
             @Override
-            public Deque<Transaction> load(String accountId)
+            public Long load(String accountId)
                     throws ResourceAccessException,
-                           DataAccessResourceFailureException  {
+                           DataAccessResourceFailureException {
                 LOGGER.fine("loaded from db");
-                Pageable request = new PageRequest(0, historyLimit);
-                return dbRepo.findForAccount(accountId,
-                                             localRoutingNum,
-                                             request);
+                Long balance = dbRepo.findBalance(accountId, localRoutingNum);
+                if (balance == null) {
+                    balance = 0L;
+                }
+                return balance;
             }
         };
         this.cache = CacheBuilder.newBuilder()
                             .maximumSize(expireSize)
-                            .expireAfterWrite(expireMinutes, TimeUnit.MINUTES)
-                            .build(load);
+                            .build(loader);
         // Initialize transaction processor.
         this.ledgerReader = reader;
         this.ledgerReader.startWithCallback(transaction -> {
@@ -139,36 +129,22 @@ public final class TransactionHistoryController {
             final String fromRouting = transaction.getFromRoutingNum();
             final String toId = transaction.getToAccountNum();
             final String toRouting = transaction.getToRoutingNum();
+            final Integer amount = transaction.getAmount();
 
             if (fromRouting.equals(localRoutingNum)
                     && this.cache.asMap().containsKey(fromId)) {
-                processTransaction(fromId, transaction);
+                Long prevBalance = cache.asMap().get(fromId);
+                this.cache.put(fromId, prevBalance - amount);
             }
             if (toRouting.equals(localRoutingNum)
                     && this.cache.asMap().containsKey(toId)) {
-                processTransaction(toId, transaction);
+                Long prevBalance = cache.asMap().get(toId);
+                this.cache.put(toId, prevBalance + amount);
             }
         });
     }
 
     /**
-     * Helper function to add a single transaction to the internal cache
-     *
-     * @param accountId   the accountId associated with the transaction
-     * @param transaction the full transaction object
-     */
-    private void processTransaction(String accountId, Transaction transaction) {
-        LOGGER.fine("modifying cache: " + accountId);
-        Deque<Transaction> tList = this.cache.asMap()
-                                             .get(accountId);
-        tList.addFirst(transaction);
-        // Drop old transactions
-        if (tList.size() > historyLimit) {
-            tList.removeLast();
-        }
-    }
-
-   /**
      * Version endpoint.
      *
      * @return  service version string
@@ -197,7 +173,7 @@ public final class TransactionHistoryController {
     @GetMapping("/healthy")
     public ResponseEntity liveness() {
         if (!ledgerReader.isAlive()) {
-            // background thread died.
+            // Background thread died.
             return new ResponseEntity<String>("Ledger reader not healthy",
                                               HttpStatus.INTERNAL_SERVER_ERROR);
         }
@@ -205,18 +181,19 @@ public final class TransactionHistoryController {
     }
 
     /**
-     * Return a list of transactions for the specified account.
+     * Return the balance for the specified account.
      *
      * The currently authenticated user must be allowed to access the account.
+     *
      * @param bearerToken  HTTP request 'Authorization' header
-     * @param accountId    the account to get transactions for.
-     * @return             a list of transactions for this account.
+     * @param accountId    the account to get the balance for
+     * @return             the balance of the account
      */
-    @GetMapping("/transactions/{accountId}")
-    public ResponseEntity<?> getTransactions(
+    @GetMapping("/balances/{accountId}")
+    public ResponseEntity<?> getBalance(
             @RequestHeader("Authorization") String bearerToken,
             @PathVariable String accountId) {
-        LOGGER.fine("request from " + accountId);
+        LOGGER.fine("request from: " + accountId);
         if (bearerToken != null && bearerToken.startsWith("Bearer ")) {
             bearerToken = bearerToken.split("Bearer ")[1];
         }
@@ -227,21 +204,9 @@ public final class TransactionHistoryController {
                 return new ResponseEntity<String>("not authorized",
                                                   HttpStatus.UNAUTHORIZED);
             }
-
             // Load from cache
-            Deque<Transaction> historyList = cache.get(accountId);
-
-            // Set artificial extra latency.
-            if (extraLatencyMillis != null) {
-                try {
-                    Thread.sleep(extraLatencyMillis);
-                } catch (InterruptedException e) {
-                    // Fake latency interrupted. Continue.
-                }
-            }
-
-            return new ResponseEntity<Collection<Transaction>>(
-                    historyList, HttpStatus.OK);
+            Long balance = cache.get(accountId);
+            return new ResponseEntity<Long>(balance, HttpStatus.OK);
         } catch (JWTVerificationException e) {
             return new ResponseEntity<String>("not authorized",
                                               HttpStatus.UNAUTHORIZED);
