+++ src/balancereader/src/main/java/anthos/samples/bankofanthos/balancereader/Transaction.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package anthos.samples.financedemo.ledgerwriter;
+package anthos.samples.bankofanthos.balancereader;
 
 import java.util.Date;
