+++ src/ledgerwriter/src/main/java/anthos/samples/bankofanthos/ledgerwriter/Transaction.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package anthos.samples.financedemo.transactionhistory;
+package anthos.samples.bankofanthos.ledgerwriter;
 
 import java.util.Date;
