+++ src/transactionhistory/src/main/java/anthos/samples/bankofanthos/transactionhistory/LedgerReader.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package anthos.samples.financedemo.transactionhistory;
+package anthos.samples.bankofanthos.transactionhistory;
 
 import java.util.logging.Logger;
