+++ src/transactionhistory/src/main/java/anthos/samples/bankofanthos/transactionhistory/TransactionRepository.java
@@ -12,10 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package anthos.samples.financedemo.balancereader;
+package anthos.samples.bankofanthos.transactionhistory;
 
+import java.util.LinkedList;
 import java.util.List;
 
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
@@ -25,21 +27,20 @@ import org.springframework.stereotype.Repository;
  */
 @Repository
 public interface TransactionRepository
-    extends CrudRepository<Transaction, Long> {
-
-    @Query(value = "SELECT "
-        + " (SELECT SUM(AMOUNT) FROM TRANSACTIONS t "
-        + "     WHERE (TO_ACCT = ?1 AND TO_ROUTE = ?2)) - "
-        + " (SELECT SUM(AMOUNT) FROM TRANSACTIONS t "
-        + "     WHERE (FROM_ACCT = ?1 AND FROM_ROUTE = ?2))",
-        nativeQuery = true)
-    Long findBalance(String accountNum, String routeNum);
-
-    @Query("SELECT t FROM Transaction t "
-        + "WHERE t.transactionId > ?1 "
-        + "ORDER BY t.transactionId ASC")
-    List<Transaction> findLatest(long latestTransaction);
+        extends CrudRepository<Transaction, Long> {
 
     @Query("SELECT MAX(transactionId) FROM Transaction")
     long latestId();
+
+    @Query("SELECT t FROM Transaction t "
+        + " WHERE (t.fromAccountNum=?1 AND t.fromRoutingNum=?2) "
+        + "   OR (t.toAccountNum=?1 AND t.toRoutingNum=?2) "
+        + " ORDER BY t.timestamp DESC")
+    LinkedList<Transaction> findForAccount(String accountNum,
+                                           String routingNum,
+                                           Pageable pager);
+
+    @Query("SELECT t FROM Transaction t "
+        + " WHERE t.transactionId > ?1 ORDER BY t.transactionId ASC")
+    List<Transaction> findLatest(long latestTransaction);
 }
